move LoanRequest -> /loans
